requirements.txt
django>=5.1,<6.0
google-generativeai>=0.8,<1.0
